Fixed #20349 Don't load test suite when not needed
https://code.djangoproject.com/ticket/20349
